Add power-assert.
