+++ package.json
@@ -19,10 +19,13 @@
     "@babel/preset-env": "^7.2.0",
     "@babel/register": "^7.0.0",
     "babel-loader": "8.0.4",
+    "babel-plugin-espower": "^3.0.0",
+    "babel-preset-power-assert": "^3.0.0",
     "chrome-webstore-manager": "^0.4.1",
     "clean-webpack-plugin": "1.0.0",
     "copy-webpack-plugin": "4.6.0",
     "mocha": "^5.2.0",
+    "power-assert": "^1.6.1",
     "puppeteer": "^1.11.0",
     "webpack": "4.27.1",
     "webpack-cli": "3.1.2",
